requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
gitpython>=3.1.40
pygithub>=2.1.1
requests>=2.31.0
pytest>=7.4.0